pyproject.toml
[tool.poetry]
name = "aws-admin"
version = "0.1.0"
description = ""
authors = ["Hugh Perkins <hughperkins@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
boto3 = "^1.35.86"
colorama = "^0.4.6"
ruamel-yaml = "^0.18.6"


[tool.poetry.group.dev.dependencies]
mypy = "^1.14.0"
types-boto3 = "^1.35.86"
types-colorama = "^0.4.15.20240311"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
